inital additional work for requesting multiple topics
