+++ routes/topics.js
@@ -1,125 +1,73 @@
 const express = require('express');
 const router = express.Router();
-const s3_helper = require('../helpers/s3_helper');
+const suggestion = require('../module/suggestion');
 
 router.get('/:topicName', async (req, res) => {
-	// Set intial variables
-	// -----
-	const topicName = req.params.topicName;
+	const result = await suggestion.topics(req.params.topicName);
+	res.status(200).json(result);
+	return;
+});
+
+router.get('/', async (req, res) => {
+	const topicsRequested = req.query.topics;
 	const data = {
-		type: 'Topic search',
-		searchToken: topicName,
+		type: 'Topic search - multiple',
+		searchToken: topicsRequested,
 		variables: {}
 	};
 
-	// Load data files
-	// -----
-	let topicClusterList;
-	let allCoocs;
-	let followUnfollowRatios;
-
-	try {
-		const clusterListDataFile = await s3_helper.getFile(
-			'topic_cluster_list.json'
-		);
-		const coocsDataFile = await s3_helper.getFile('allCoocs.json');
-		const followUnfollowRatiosDataFile = await s3_helper.getFile(
-			'topic_ratios.json'
-		);
-
-		topicClusterList = JSON.parse(clusterListDataFile.Body.toString());
-		allCoocs = JSON.parse(coocsDataFile.Body.toString());
-		followUnfollowRatios = JSON.parse(
-			followUnfollowRatiosDataFile.Body.toString()
-		);
-	} catch (err) {
-		console.log(err);
-		data.error = err;
-		json.status(200).json(data);
-		return;
-	}
-
-	// Find topic in list
-	// -----
-	const foundTopic = topicClusterList.filter((entry) =>
-		entry.name === topicName ? entry : null
-	);
-
-	if (foundTopic.length <= 0) {
-		data.error = 'Topic not found';
+	if (!topicsRequested) {
+		data.error = 'Topic query parameter not defined';
 		res.status(200).json(data);
 		return;
 	}
 
-	const foundTopicSingle = foundTopic[0];
-	const foundTopicClusterNo = foundTopicSingle.cluster_group;
-	data.clusterNumber = foundTopicClusterNo;
+	const topics = req.query.topics.split(',');
+	const topicPromises = [];
+	topics.map((topic) => {
+		new Promise(function(resolve, reject) {
+			topicPromises.push(suggestion.topics(topic));
+		});
+	});
 
-	// Get all the other topics with the matching group
-	// -----
-	const allTopicsInCluster = topicClusterList
-		.filter(function(entry) {
-			if (
-				entry.cluster_group === foundTopicClusterNo &&
-				entry.name !== topicName
-			) {
-				return entry;
-			}
+	Promise.all(topicPromises)
+		.then((results) => {
+			data.results = results;
+
+			// also return combined results ordered by occurence (how many times it came up for seperate topics) and ratio
+
+			const combined_sorted_topic_results = results
+				.filter((entry) => {
+					if (entry.variables.nonMatchingTopicsSorted) {
+						return entry;
+					}
+				})
+				.map((entry) => {
+					return entry.variables.nonMatchingTopicsSorted;
+				})
+				.sort((a, b) => (a.ratio < b.ratio ? 1 : -1));
+
+			const mergedResults = [].concat.apply(
+				[],
+				combined_sorted_topic_results
+			);
+
+			data.nonMatchingTopicsSortedCombined = mergedResults;
+
+			const nonMatchingTopicsSortedCombinedClean = mergedResults.map(
+				(entry) => entry.concept_name
+			);
+			data.nonMatchingTopicsSortedCombinedClean = nonMatchingTopicsSortedCombinedClean;
+
+			res.status(200).json(data);
+			return;
 		})
-		.map(function(entry) {
-			return {
-				name: entry.name,
-				cluster_group: entry.cluster_group
-			};
+		.catch((error) => {
+			data.error = error;
+			res.status(200).json(data);
+			return;
 		});
 
-	data.variables.allTopicsInCluster = allTopicsInCluster;
-
-	// Get the list of correlated content
-	// -----
-	const correlatedTopics = allCoocs[`topics:${topicName}`];
-	data.variables.correlatedTopics = correlatedTopics;
-
-	// Find the remaining topics that are not related at all
-	// -----
-	const allTopicsInClusterArr = allTopicsInCluster.map((entry) => entry.name);
-	data.variables.allTopicsInClusterArr = allTopicsInClusterArr;
-
-	const correlatedTopicsArr = Object.keys(correlatedTopics).map((entry) =>
-		entry.replace('topics:', '')
-	);
-	data.variables.correlatedTopicsArr = correlatedTopicsArr;
-
-	const nonMatchingTopics = allTopicsInClusterArr.filter((entry) =>
-		!correlatedTopicsArr.includes(entry) ? entry : null
-	);
-	data.variables.nonMatchingTopics = nonMatchingTopics;
-
-	if (nonMatchingTopics.length <= 0) {
-		data.error = 'All topics match - no outliers found';
-		res.status(200).json(data);
-		return;
-	}
-
-	// Sort by follow/unfollow ratio
-	// -----
-	const nonMatchingTopicsSorted = followUnfollowRatios
-		.filter((entry) =>
-			nonMatchingTopics.includes(entry.concept_name) ? entry : null
-		)
-		.sort((a, b) => (a.ratio < b.ratio ? 1 : -1));
-	data.variables.nonMatchingTopicsSorted = nonMatchingTopicsSorted;
-
-	// Clean the final result to return only the topic names sorted
-	// -----
-	const nonMatchingTopicsSortedClean = nonMatchingTopicsSorted.map(
-		(entry) => entry.concept_name
-	);
-	data.nonMatchingTopicsSortedClean = nonMatchingTopicsSortedClean;
-
-	// Return results
-	// -----
-	res.status(200).json(data);
 	return;
 });
