+++ views/index.hbs
@@ -28,12 +28,17 @@
   </h2>
   <ul>
     <li>
-      <a href="/topic/Brexit">Get suggestions for the topic <b>Brexit</b></a>
+      <a href="/topics/Brexit">Get suggestions for the topic <b>Brexit</b></a>
     </li>
     <li>
-      <a href="/topic/Blockchain">Get suggestions for the topic <b>Blockchain</b></a>
+      <a href="/topics/Blockchain">Get suggestions for the topic <b>Blockchain</b></a>
+    </li>
+    <li>
+      <a href="/topics/?topics=Private%20equity,Investment%20Banking,Hedge%20funds">Get suggestions for the topics
+        <b>Private equity, Investment Banking & Hedge funds</b></a>
     </li>
   </ul>
+
   <script src="/static/js/script.js"></script>
 </body>
